Shell collapse: hoist stateful chart-render helpers to Ui.ChartRenderHelper
The ScottPlot legend-panel tracking (_legendPanels + ClearChart/ShowChartLegend) and the
empty-state "No Data" chart (RefreshEmptyChart) were duplicated verbatim in Lite
ServerTab.Charts.cs and Darling ViewerServerTab.ChartHelpers.cs. Collapsed into a shared
PerformanceMonitor.Ui.ChartRenderHelper; each app holds one instance and forwards its many
ClearChart/RefreshEmptyChart/ShowChartLegend call sites through thin per-app methods (call
sites unchanged). The Query Heatmap's direct _legendPanels write goes through a new
SetLegendPanel accessor. The stateless theme/axis/y-limit forwarders stay per-app (already
delegate to the shared ChartStyle). Net -116 lines. Prerequisite for sharing the per-tab
chart bodies (System Events counter charts next).

The deprecated Full Dashboard keeps its own copies.

Verified: all three apps build, both capability pins green (Lite 6/6, Darling 6/6).
